re-record cassettee after rename
