+++ lib/easypost/models.rb
@@ -29,7 +29,7 @@ require_relative 'models/refund'
 require_relative 'models/report'
 require_relative 'models/scan_form'
 require_relative 'models/shipment'
-require_relative 'models/smartrate'
+require_relative 'models/smart_rate'
 require_relative 'models/tax_identifier'
 require_relative 'models/tracker'
 require_relative 'models/user'
