+++ lib/easypost/models/smart_rate.rb
