+++ spec/cassettes/smart_rate/EasyPost_Services_SmartRate_deliver_by_retrieve_the_estimated_delivery_date.yml
@@ -35,7 +35,7 @@ http_interactions:
       Referrer-Policy:
       - strict-origin-when-cross-origin
       X-Ep-Request-Uuid:
-      - b8e53c2366902919e79b13ec00082fef
+      - b8e53c1c669056cbe7873cbe0041fafa
       Cache-Control:
       - private, no-cache, no-store
       Pragma:
@@ -45,22 +45,22 @@ http_interactions:
       Content-Type:
       - application/json; charset=utf-8
       X-Runtime:
-      - '0.077908'
+      - '0.049003'
       Transfer-Encoding:
       - chunked
       X-Node:
-      - bigweb34nuq
+      - bigweb40nuq
       X-Version-Label:
-      - easypost-202407111640-83b200930f-master
+      - easypost-202407111940-3cb16a3462-master
       X-Backend:
       - easypost
       X-Proxied:
       - extlb2nuq fa152d4755
-      - intlb3nuq fa152d4755
+      - intlb4nuq fa152d4755
       Strict-Transport-Security:
       - max-age=31536000; includeSubDomains; preload
     body:
       encoding: UTF-8
       string: '{"carriers_without_tint_estimates":null,"from_zip":"94107","planned_ship_date":"2024-07-16","results":[{"carrier":"USPS","easypost_time_in_transit_data":{"days_in_transit":{"percentile_50":1,"percentile_75":2,"percentile_85":2,"percentile_90":3,"percentile_95":3,"percentile_97":4,"percentile_99":6},"easypost_estimated_delivery_date":"2024-07-17"},"service":"express"},{"carrier":"USPS","easypost_time_in_transit_data":{"days_in_transit":{"percentile_50":2,"percentile_75":3,"percentile_85":3,"percentile_90":4,"percentile_95":6,"percentile_97":6,"percentile_99":9},"easypost_estimated_delivery_date":"2024-07-18"},"service":"groundadvantage"},{"carrier":"USPS","easypost_time_in_transit_data":{"days_in_transit":{"percentile_50":3,"percentile_75":3,"percentile_85":3,"percentile_90":6,"percentile_95":6,"percentile_97":6,"percentile_99":9},"easypost_estimated_delivery_date":"2024-07-18"},"service":"librarymail"},{"carrier":"USPS","easypost_time_in_transit_data":{"days_in_transit":{"percentile_50":3,"percentile_75":3,"percentile_85":3,"percentile_90":4,"percentile_95":6,"percentile_97":6,"percentile_99":9},"easypost_estimated_delivery_date":"2024-07-18"},"service":"mediamail"},{"carrier":"USPS","easypost_time_in_transit_data":{"days_in_transit":{"percentile_50":2,"percentile_75":3,"percentile_85":3,"percentile_90":3,"percentile_95":4,"percentile_97":6,"percentile_99":7},"easypost_estimated_delivery_date":"2024-07-18"},"service":"priority"}],"saturday_delivery":null,"to_zip":"90277"}'
-  recorded_at: Thu, 11 Jul 2024 18:48:57 GMT
+  recorded_at: Thu, 11 Jul 2024 22:03:55 GMT
 recorded_with: VCR 6.1.0
