+++ spec/cassettes/smart_rate/EasyPost_Services_SmartRate_deliver_on_retrieve_a_recommended_ship_date.yml
@@ -35,7 +35,7 @@ http_interactions:
       Referrer-Policy:
       - strict-origin-when-cross-origin
       X-Ep-Request-Uuid:
-      - b8e53c1d66902a39e779a00d000947e8
+      - b8e53c21669056cce7873cc50041fbdb
       Cache-Control:
       - private, no-cache, no-store
       Pragma:
@@ -45,22 +45,22 @@ http_interactions:
       Content-Type:
       - application/json; charset=utf-8
       X-Runtime:
-      - '0.069828'
+      - '0.054759'
       Transfer-Encoding:
       - chunked
       X-Node:
-      - bigweb42nuq
+      - bigweb53nuq
       X-Version-Label:
-      - easypost-202407111640-83b200930f-master
+      - easypost-202407111940-3cb16a3462-master
       X-Backend:
       - easypost
       X-Proxied:
       - extlb2nuq fa152d4755
-      - intlb3nuq fa152d4755
+      - intlb4nuq fa152d4755
       Strict-Transport-Security:
       - max-age=31536000; includeSubDomains; preload
     body:
       encoding: UTF-8
       string: '{"carriers_without_tint_estimates":null,"desired_delivery_date":"2024-07-16","from_zip":"94107","results":[{"carrier":"USPS","easypost_time_in_transit_data":{"days_in_transit":{"percentile_50":2,"percentile_75":2,"percentile_85":2,"percentile_90":3,"percentile_95":3,"percentile_97":4,"percentile_99":7},"delivery_date_confidence":0.47,"estimated_transit_days":1,"ship_on_date":"2024-07-15"},"service":"express"},{"carrier":"USPS","easypost_time_in_transit_data":{"days_in_transit":{"percentile_50":2,"percentile_75":3,"percentile_85":4,"percentile_90":4,"percentile_95":5,"percentile_97":5,"percentile_99":9},"delivery_date_confidence":0.21,"estimated_transit_days":3,"ship_on_date":"2024-07-13"},"service":"groundadvantage"},{"carrier":"USPS","easypost_time_in_transit_data":{"days_in_transit":{"percentile_50":2,"percentile_75":3,"percentile_85":4,"percentile_90":4,"percentile_95":5,"percentile_97":6,"percentile_99":9},"delivery_date_confidence":0.25,"estimated_transit_days":3,"ship_on_date":"2024-07-13"},"service":"librarymail"},{"carrier":"USPS","easypost_time_in_transit_data":{"days_in_transit":{"percentile_50":2,"percentile_75":3,"percentile_85":3,"percentile_90":4,"percentile_95":5,"percentile_97":5,"percentile_99":7},"delivery_date_confidence":0.24,"estimated_transit_days":3,"ship_on_date":"2024-07-13"},"service":"mediamail"},{"carrier":"USPS","easypost_time_in_transit_data":{"days_in_transit":{"percentile_50":2,"percentile_75":3,"percentile_85":4,"percentile_90":4,"percentile_95":5,"percentile_97":5,"percentile_99":7},"delivery_date_confidence":0.19,"estimated_transit_days":3,"ship_on_date":"2024-07-13"},"service":"priority"}],"saturday_delivery":null,"to_zip":"90277"}'
-  recorded_at: Thu, 11 Jul 2024 18:53:45 GMT
+  recorded_at: Thu, 11 Jul 2024 22:03:56 GMT
 recorded_with: VCR 6.1.0
