+++ spec/smart_rate_spec.rb
